+++ src/lua/graphic/rect.odin
@@ -81,7 +81,7 @@ RECT_FUNCTION :: lua_common.LuaFunction {
 			shader_args = shader_args,
 		}
 
-		core.add_group_to_render_queue(i32(zIndex), texture_path, u64(shader), fixed, tiled, props)
+		core.add_group_to_render_queue(i32(zIndex), texture_path, u32(shader), fixed, tiled, props)
 
 		return 0
 	},
